requirements.txt
a2wsgi==1.10.10
aiosmtplib==4.0.2
aiosqlite==0.21.0
alembic==1.17.0
annotated-types==0.7.0
anyio==4.8.0
apache-airflow==3.1.0
apache-airflow-core==3.1.0
apache-airflow-providers-common-compat==1.7.4
apache-airflow-providers-common-io==1.6.3
apache-airflow-providers-common-sql==1.28.1
apache-airflow-providers-smtp==2.3.1
apache-airflow-providers-standard==1.9.0
apache-airflow-task-sdk==1.1.0
argcomplete==3.6.2
asgiref==3.8.1
attrs==25.4.0
audioop-lts==0.2.1
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
cadwyn==5.6.0
certifi==2025.1.31
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
comtypes==1.4.10
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==46.0.2
Deprecated==1.2.18
dill==0.4.0
distlib==0.4.0
distro==1.9.0
dj-database-url==3.0.1
Django==5.1.6
django-cors-headers==4.9.0
djangorestframework==3.15.2
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.119.0
fastapi-cli==0.0.13
filelock==3.20.0
Flask==3.1.0
flask-cors==5.0.1
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.176.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
greenback==1.2.1
greenlet==3.2.4
grpcio==1.71.0
grpcio-status==1.71.0
gspread==6.2.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.7.1
httpx==0.28.1
identify==2.6.15
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
libcst==1.8.5
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.8.0
msgspec==0.19.0
natsort==8.4.0
nodeenv==1.9.1
numpy==2.2.5
oauthlib==3.3.1
openai==1.63.0
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
outcome==1.3.0.post0
packaging==25.0
pathlib_abc==0.5.1
pathspec==0.12.1
pendulum==3.1.0
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.3.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.1.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
pygame==2.6.1
Pygments==2.19.2
pygtrie==2.5.0
PyJWT==2.10.1
pyparsing==3.2.1
pypi==2.1
pypiwin32==223
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.20
python-slugify==8.0.4
python-socketio==5.12.1
pyttsx3==2.98
pytz==2025.2
pywin32==308
PyYAML==6.0.3
PyYAML-ft==8.0.0
referencing==0.37.0
requests==2.32.3
requests-oauthlib==2.0.0
retryhttp==1.3.3
rich==14.2.0
rich-argparse==1.7.1
rich-toolkit==0.15.1
rpds-py==0.27.1
rsa==4.9
scikit-learn==1.7.2
scipy==1.16.3
SDK==1.0.0
setproctitle==1.3.7
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.14.1
speechSynthesis==1.0
SQLAlchemy==2.0.44
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.0
sqlparse==0.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.46.1
structlog==25.4.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
threadpoolctl==3.6.0
tqdm==4.67.1
typer==0.19.2
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
uc-micro-py==1.0.3
universal_pathlib==0.3.3
uritemplate==4.1.1
urllib3==2.3.0
uuid6==2025.0.1
uvicorn==0.34.0
virtualenv==20.35.3
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.11.0
windows==0.1.1
wirerope==1.0.0
wrapt==1.17.3
wsproto==1.2.0
zipp==3.23.0
